pyproject.toml
[tool.poetry]
name = "heater"
version = "0.0.0"
readme = "README.md"
description = "heater makes heatmaps."
authors = ["Bogdan Popa <popa.bogdanp@gmail.com>"]

[tool.poetry.dependencies]
cycler = "^0.10.0"
matplotlib = "^2.2"
numpy = "^1.14"
pillow = "^5.1"
pyparsing = "^2.2"
python = "^3.5"
python-dateutil = "^2.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.5"
flake8-bugbear = "^18.2"
flake8-quotes = "^1.0"
isort = "^4.3"
pytest = "^3.5"
pytest-cov = "^2.5"
bumpversion = "^0.5.3"
